Refactor allocator module
Decouple Solana-only code - step 1

Less #[cfg(target_os = "solana")] and #[cfg(not(target_os = "solana"))]

Extract solana_allocator.rs module

Extract state_account_allocator.rs

Cleanup imports

Add useful comments

Promote allocator module to directory

Move solana_allocator.rs and state_account_allocator.rs into allocator.rs

Extract Alloc trait

Use acc_allocator() and make STATE_ACCOUNT_ALLOCATOR private

Remove static mut STATE_ACCOUNT_ALLOCATOR

Extract PROGRAM_DATA_INPUT_PARAMETERS_OFFSET constant

Extract impl_global_alloc macro

Extract solana module

Extract Alloc::heap method

Inline heap methods
